refactor: file rename
This commit also changes helper mssg in RegistrationPage
